@@ -11,30 +11,26 @@ description: A to do list flutter project.
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.0+1
+version: 1.0.1+5
 
 environment:
-  sdk: ">=2.5.2 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  # cupertino_icons: ^0.1.2
   english_words: ^4.0.0
-  flutter_login: ^1.0.14
   shared_preferences: ^2.0.6
   http: ^0.13.3
   rxdart: ^0.27.1
   form_field_validator: ^1.0.1
   flutter_keyboard_size: ^1.0.0+3
   equatable: ^2.0.2
-  circular_check_box: ^1.0.4
   splashscreen: ^1.3.5
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.3
-  circular_menu: ^1.2.2
 
 dev_dependencies:
   flutter_test:
