@@ -1,8 +1,11 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_keyboard_size/flutter_keyboard_size.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 import 'package:todolist/UI/title_card.dart';
 import 'package:todolist/bloc/blocs/user_bloc_provider.dart';
 import 'package:todolist/models/global.dart';
+import 'package:todolist/models/group.dart';
 import 'package:todolist/models/subtasks.dart';
 import 'package:todolist/models/tasks.dart';
 import 'package:todolist/widgets/global_widgets/background_color_container.dart';
@@ -10,60 +13,75 @@ import 'package:todolist/widgets/task_widgets/add_subtask_widget.dart';
 import 'package:todolist/widgets/task_widgets/subtask_list_item_widget.dart';
 
 class SubtaskListTab extends StatefulWidget {
-  final Task task;
-
-  SubtaskListTab({@required this.task});
-
+  static const routeName = '/listSubtasksTab';
   @override
   _SubtaskListTabState createState() => _SubtaskListTabState();
 }
 
 class _SubtaskListTabState extends State<SubtaskListTab> {
   //List<Subtask> subtasks;
-  SubtaskBloc subtaskBloc;
-  int orderBy;
+  late SubtaskBloc subtaskBloc;
+  late Group group;
+  late Task task;
+  late double unitHeightValue;
+  late String orderBy;
+  bool reorder;
 
-  @override
-  void initState() {
-    subtaskBloc = SubtaskBloc(widget.task.taskKey);
-    orderBy = 0;
-    super.initState();
+  _SubtaskListTabState() : reorder = false {
+    getOrderBy();
   }
 
   @override
   Widget build(BuildContext context) {
+    final args =
+        ModalRoute.of(context)!.settings.arguments as SubtaskListTabArguments;
+    task = args.task;
+    group = args.group;
+    unitHeightValue = MediaQuery.of(context).size.height * 0.001;
+    subtaskBloc = SubtaskBloc(task);
     return KeyboardSizeProvider(
       child: SafeArea(
-        child: Scaffold(
-          appBar: AppBar(
-            title: Text(
-              widget.task.title,
-              style: appTitleStyle,
-            ),
-            centerTitle: true,
-            backgroundColor: Colors.white,
-            elevation: 0,
-            leading: IconButton(
-              icon: Icon(Icons.arrow_back, size: 32.0, color: darkBlueGradient),
-              onPressed: () {
-                Navigator.pop(context);
-              },
-              color: Colors.blueGrey,
-            ),
-            actions: [_popupMenuButton()],
-          ),
-          body: Stack(
-            children: <Widget>[
-              BackgroundColorContainer(
-                startColor: lightGreenBlue,
-                endColor: darkGreenBlue,
-                widget: TitleCard(title: 'To Do', child: _buildStreamBuilder()),
+        child: GestureDetector(
+          onTap: () {
+            FocusScopeNode currentFocus = FocusScope.of(context);
+
+            if (!currentFocus.hasPrimaryFocus) {
+              currentFocus.unfocus();
+            }
+          },
+          child: Scaffold(
+            appBar: AppBar(
+              title: Text(
+                task.title,
+                style: appTitleStyle(unitHeightValue),
               ),
-              AddSubtask(
-                length: widget.task.subtasks.length,
-                subtaskBloc: subtaskBloc,
+              centerTitle: true,
+              backgroundColor: Colors.white,
+              elevation: 0,
+              leading: IconButton(
+                icon: Icon(Icons.arrow_back,
+                    size: 32.0 * unitHeightValue, color: darkBlueGradient),
+                onPressed: () {
+                  Navigator.pop(context);
+                },
+                color: Colors.blueGrey,
               ),
-            ],
+              actions: [_popupMenuButton(), SizedBox(width: 10)],
+            ),
+            body: Stack(
+              children: <Widget>[
+                BackgroundColorContainer(
+                  startColor: lightGreenBlue,
+                  endColor: darkGreenBlue,
+                  widget:
+                      TitleCard(title: 'To Do', child: _buildStreamBuilder()),
+                ),
+                AddSubtask(
+                  length: task.subtasks.length,
+                  subtaskBloc: subtaskBloc,
+                ),
+              ],
+            ),
           ),
         ),
       ),
@@ -74,46 +92,29 @@ class _SubtaskListTabState extends State<SubtaskListTab> {
     return StreamBuilder(
       // Wrap our widget with a StreamBuilder
       stream: subtaskBloc.getSubtasks, // pass our Stream getter here
-      initialData: widget.task.subtasks, // provide an initial data
+      initialData: task.subtasks, // provide an initial data
       builder: (context, snapshot) {
         switch (snapshot.connectionState) {
           case ConnectionState.none:
-            return Container(
-              child: Center(
-                child: Text("No Connection Message"),
-              ),
-            );
+            break;
           case ConnectionState.active:
             print("Active Data: " +
                 snapshot.data.toString() +
                 " @" +
                 DateTime.now().toString());
-            if (snapshot.data.isNotEmpty) {
-              widget.task.subtasks = snapshot.data;
-              print("Group Task List: " +
-                  widget.task.subtasks.toString() +
-                  " @" +
-                  DateTime.now().toString());
-
-              return _buildList();
+            if (snapshot.hasData && !listEquals(task.subtasks, snapshot.data)) {
+              task.subtasks = snapshot.data!;
             }
-            return SizedBox.shrink();
-            break;
+            if (reorder) {
+              reorder = false;
+            }
+            return _buildList();
           case ConnectionState.waiting:
-            print("Waiting Data" +
-                snapshot.data.toString() +
-                " @" +
-                DateTime.now().toString());
-            return Center(child: CircularProgressIndicator());
-            break;
+            return Center(
+                child: CircularProgressIndicator(color: Colors.black54));
           case ConnectionState.done:
             print("Done Data: " + snapshot.toString());
-            if (snapshot.data.isNotEmpty) {
-              widget.task.subtasks = snapshot.data;
-
-              return _buildList();
-            }
-            return SizedBox.shrink();
+            break;
         }
         return CircularProgressIndicator();
       },
@@ -121,27 +122,29 @@ class _SubtaskListTabState extends State<SubtaskListTab> {
   }
 
   Widget _buildList() {
-    _orderBy(orderBy);
+    _orderBy();
     return Theme(
       data: ThemeData(canvasColor: Colors.transparent),
       key: UniqueKey(),
       child: ListView(
         key: UniqueKey(),
         padding: EdgeInsets.only(top: 175, bottom: 90),
-        children: widget.task.subtasks.map<Dismissible>((Subtask item) {
+        children: task.subtasks.map<Dismissible>((Subtask item) {
           return _buildListTile(item);
         }).toList(),
       ),
     );
   }
 
-  Widget _buildListTile(Subtask subtask) {
+  Dismissible _buildListTile(Subtask subtask) {
     return Dismissible(
       key: Key(subtask.subtaskKey),
       child: ListTile(
         key: Key(subtask.title),
         title: SubtaskListItemWidget(
           subtask: subtask,
+          subtaskBloc: subtaskBloc,
+          group: group,
         ),
       ),
       background: Container(
@@ -150,6 +153,7 @@ class _SubtaskListTabState extends State<SubtaskListTab> {
         child: Icon(
           Icons.delete,
           color: lightGreenBlue,
+          size: 28 * unitHeightValue,
         ),
       ),
       onDismissed: (direction) {
@@ -169,16 +173,15 @@ class _SubtaskListTabState extends State<SubtaskListTab> {
   }
 
   void removeSubtask(Subtask subtask) {
-    if (widget.task.subtasks.contains(subtask)) {
+    if (task.subtasks.contains(subtask)) {
       setState(() {
-        widget.task.subtasks.remove(subtask);
+        task.subtasks.remove(subtask);
       });
     }
   }
 
   void reAddSubtask(Subtask subtask) async {
-    await subtaskBloc.addSubtask(
-        subtask.title, subtask.index, subtask.completed);
+    await subtaskBloc.addSubtask(subtask.title);
     setState(() {});
   }
 
@@ -188,9 +191,9 @@ class _SubtaskListTabState extends State<SubtaskListTab> {
   }
 
   PopupMenuButton _popupMenuButton() {
-    return PopupMenuButton<int>(
-      icon: Icon(Icons.sort, size: 32.0, color: darkBlueGradient),
-      iconSize: 24.0,
+    return PopupMenuButton<String>(
+      icon: Icon(Icons.sort,
+          size: 32.0 * unitHeightValue, color: darkBlueGradient),
       color: darkGreenBlue,
       offset: Offset(0, 50),
       shape: RoundedRectangleBorder(
@@ -199,41 +202,57 @@ class _SubtaskListTabState extends State<SubtaskListTab> {
         ),
       ),
       onSelected: (value) {
+        saveOrderBy(value);
         setState(() {
-          orderBy = value;
+          reorder = true;
         });
       },
       itemBuilder: (context) => [
-        PopupMenuItem<int>(
-          value: 0,
+        PopupMenuItem<String>(
+          value: "Alphabetical",
           child: Row(children: [
-            Icon(Icons.sort_by_alpha),
+            Icon(Icons.sort_by_alpha, size: 24 * unitHeightValue),
             SizedBox(width: 6.0),
             Text(
               "Alphabetical",
-              style: TextStyle(color: Colors.white),
+              style: TextStyle(
+                  color: Colors.white, fontSize: 18 * unitHeightValue),
             )
           ]),
         ),
-        PopupMenuItem<int>(
-          value: 1,
+        PopupMenuItem<String>(
+          value: "Recent-Oldest",
           child: Row(children: [
-            Icon(Icons.date_range),
+            Icon(Icons.date_range, size: 24 * unitHeightValue),
             SizedBox(width: 6.0),
             Text(
               "Recent-Oldest",
-              style: TextStyle(color: Colors.white),
+              style: TextStyle(
+                  color: Colors.white, fontSize: 18 * unitHeightValue),
             )
           ]),
         ),
-        PopupMenuItem<int>(
-          value: 2,
+        PopupMenuItem<String>(
+          value: "Oldest-Recent",
           child: Row(children: [
-            Icon(Icons.date_range),
+            Icon(Icons.date_range, size: 24 * unitHeightValue),
             SizedBox(width: 6.0),
             Text(
               "Oldest-Recent",
-              style: TextStyle(color: Colors.white),
+              style: TextStyle(
+                  color: Colors.white, fontSize: 18 * unitHeightValue),
+            )
+          ]),
+        ),
+        PopupMenuItem<String>(
+          value: "Due Date",
+          child: Row(children: [
+            Icon(Icons.date_range, size: 24 * unitHeightValue),
+            SizedBox(width: 6.0),
+            Text(
+              "Due Date",
+              style: TextStyle(
+                  color: Colors.white, fontSize: 18 * unitHeightValue),
             )
           ]),
         ),
@@ -241,22 +260,41 @@ class _SubtaskListTabState extends State<SubtaskListTab> {
     );
   }
 
-  _orderBy(int value) {
-    orderBy = value;
-    switch (value) {
-      case 0:
-        widget.task.subtasks.sort(
+  _orderBy() {
+    switch (orderBy) {
+      case "Alphabetical":
+        task.subtasks.sort(
             (a, b) => a.title.toLowerCase().compareTo(b.title.toLowerCase()));
         break;
-      case 1:
-        widget.task.subtasks
-            .sort((a, b) => b.timeUpdated.compareTo(a.timeUpdated));
+      case "Recent-Oldest":
+        task.subtasks.sort((a, b) => b.timeCreated.compareTo(a.timeCreated));
         break;
-      case 2:
-        widget.task.subtasks
-            .sort((a, b) => a.timeUpdated.compareTo(b.timeUpdated));
+      case "Oldest-Recent":
+        task.subtasks.sort((a, b) => a.timeCreated.compareTo(b.timeCreated));
+        break;
+      case "Due Date":
+        task.subtasks.sort((a, b) => a.deadline.compareTo(b.deadline));
         break;
       default:
     }
   }
+
+  /// Get order list from persistant storage.
+  void getOrderBy() async {
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+    this.orderBy = prefs.getString('SUBTASK_ORDER_LIST') ?? "Recent-Oldest";
+  }
+
+  /// Save orderlist to Device's persistant storage
+  Future<void> saveOrderBy(String orderBy) async {
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+    await prefs.setString('SUBTASK_ORDER_LIST', orderBy);
+    this.orderBy = orderBy;
+  }
+}
+
+class SubtaskListTabArguments {
+  final Group group;
+  final Task task;
+  SubtaskListTabArguments(this.group, this.task);
 }
