@@ -1,33 +1,37 @@
 import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
+import 'package:todolist/models/global.dart';
 
 // ignore: must_be_immutable
 class GroupMember extends Equatable {
   /// Member's First name
-  final String firstname;
+  late String firstname;
 
   /// Member's Last name
-  final String lastname;
+  late String lastname;
 
   /// Member's Username
-  final String username;
+  late String username;
 
   /// Member's Email Address
-  final String emailaddress;
+  late String emailaddress;
 
   /// Member's Phone Number
-  final String phonenumber;
+  late String phonenumber;
 
   /// Avatar Image
-  final ImageProvider avatar;
+  late var avatar;
 
+  bool selectedForAssignment = false;
+
+  GroupMember.blank();
   GroupMember(
-      {this.firstname,
-      this.lastname,
-      this.emailaddress,
-      this.username,
-      this.phonenumber,
-      this.avatar});
+      {required this.firstname,
+      required this.lastname,
+      required this.emailaddress,
+      required this.username,
+      required this.phonenumber,
+      required this.avatar});
 
   factory GroupMember.fromJson(Map<String, dynamic> parsedJson) {
     return GroupMember(
@@ -47,25 +51,31 @@ class GroupMember extends Equatable {
   String initials() => firstname[0].toUpperCase() + lastname[0].toUpperCase();
 
   /// Create CircleAvatar
-  CircleAvatar cAvatar({double radius = 12}) {
+  CircleAvatar cAvatar({
+    double radius = 12,
+    Color color = darkBlue,
+    required double unitHeightValue,
+  }) {
     return CircleAvatar(
       //backgroundImage: member.avatar,
-      //backgroundColor: Colors.grey,
-      child: Text(
-        this.initials(),
-        style: TextStyle(
-          fontFamily: "Segoe UI",
-          fontWeight: FontWeight.bold,
-          fontSize: radius + 2,
-          color: Colors.white,
-        ),
-      ),
-      radius: radius,
+      backgroundColor: color,
+      child: selectedForAssignment
+          ? Icon(Icons.check, size: radius * unitHeightValue + 8, color: Colors.white)
+          : Text(
+              this.initials(),
+              style: TextStyle(
+                fontFamily: "Segoe UI",
+                fontWeight: FontWeight.bold,
+                fontSize: radius * unitHeightValue + 2,
+                color: Colors.white,
+              ),
+            ),
+      radius: radius * unitHeightValue,
     );
   }
 
   @override
   String toString() {
-    return "Member: $username";
+    return "Member: $username Assigned: $selectedForAssignment";
   }
 }
