@@ -1,18 +1,17 @@
-class Subtask {
+import 'package:equatable/equatable.dart';
+
+import 'groupmember.dart';
+
+// ignore: must_be_immutable
+class Subtask extends Equatable {
   /// Subtask Name/Title
   String title;
 
   /// Subtask ID
   int subtaskId;
 
-  /// Subtask Index in the list
-  int index;
-
   /// Subtask Key
   String subtaskKey;
-
-  /// Group Name
-  String group;
 
   /// Has the subtask been completed
   bool completed;
@@ -23,31 +22,37 @@ class Subtask {
   /// Time Updated
   DateTime timeUpdated;
 
+  DateTime now = DateTime.now();
+
+  /// Deadline
+  late DateTime deadline;
+
   /// Not Implemented
   String note;
 
-  /// Not Implemented
-  String repeats;
+  List<GroupMember> assignedTo = [];
 
-  /// Not Implemented
-  List<DateTime> reminders;
-  //DateTime deadline;
+  late List<GroupMember> allGroupMembers;
 
-  Subtask(this.title, this.group, this.completed, this.subtaskId, this.note,
-      this.subtaskKey, this.index, this.timeCreated, this.timeUpdated);
+  Subtask(this.title, this.completed, this.subtaskId, this.note,
+      this.subtaskKey, this.timeCreated, this.timeUpdated) {
+    deadline = DateTime(now.year, 12, 31);
+  }
 
   factory Subtask.fromJson(Map<String, dynamic> parsedJson) {
     return Subtask(
-        parsedJson['title'],
-        parsedJson['group'],
-        parsedJson['completed'],
-        parsedJson['id'],
-        parsedJson['note'],
-        parsedJson['subtask_key'],
-        parsedJson['index'],
-        DateTime.parse(parsedJson['time_created']),
-        DateTime.parse(parsedJson['time_updated']));
+      parsedJson['title'],
+      parsedJson['completed'],
+      parsedJson['id'],
+      parsedJson['note'],
+      parsedJson['subtask_key'],
+      DateTime.parse(parsedJson['time_created']),
+      DateTime.parse(parsedJson['time_updated']),
+    );
   }
+
+  @override
+  List<Object> get props => [subtaskKey];
 
   @override
   String toString() {
