@@ -4,9 +4,9 @@ import 'package:todolist/bloc/blocs/user_bloc_provider.dart';
 
 class AddTask extends StatefulWidget {
   const AddTask({
-    Key key,
-    @required this.length,
-    @required this.taskbloc,
+    Key? key,
+    required this.length,
+    required this.taskbloc,
   }) : super(key: key);
 
   final int length;
@@ -19,22 +19,27 @@ class AddTask extends StatefulWidget {
 class _AddTaskState extends State<AddTask> {
   TextEditingController controller = new TextEditingController();
   FocusNode textfieldFocus = new FocusNode();
-
-  double height = 60.0;
-  double width = 250.0;
-  double bottom = 0;
-  double focusWidth, marginH;
-  Size size;
+  late double unitValueHeight,
+      unitValueWidth,
+      height,
+      defaultWidth,
+      focusWidth,
+      marginH;
+  late Size size;
 
   @override
   Widget build(BuildContext context) {
     size = MediaQuery.of(context).size;
-    focusWidth = size.width * 0.9;
-    width = (textfieldFocus.hasPrimaryFocus) ? focusWidth : 250;
+    unitValueHeight = size.height * 0.001;
+    unitValueWidth = size.width * 0.001;
+    focusWidth = unitValueWidth * 900;
+    defaultWidth = unitValueWidth * 250;
+    height = unitValueHeight * 60;
+    double width = (textfieldFocus.hasPrimaryFocus) ? focusWidth : defaultWidth;
     marginH = (size.width - width) / 2;
     return Consumer<ScreenHeight>(builder: (context, _res, child) {
       return Positioned(
-        bottom: bottom,
+        bottom: 0,
         child: Container(
           margin: EdgeInsets.symmetric(horizontal: marginH, vertical: 25),
           decoration: BoxDecoration(
@@ -61,12 +66,9 @@ class _AddTaskState extends State<AddTask> {
                   textAlignVertical: TextAlignVertical.center,
                   keyboardType: TextInputType.text,
                   textInputAction: TextInputAction.go,
+                  style: TextStyle(fontSize: 16 * unitValueHeight),
                   onTap: () {
-                    if (width != focusWidth) {
-                      setState(() {
-                        width = focusWidth;
-                      });
-                    }
+                    setState(() {});
                   },
                   onSubmitted: (_) {
                     setState(() {
@@ -86,7 +88,8 @@ class _AddTaskState extends State<AddTask> {
                     //contentPadding: EdgeInsets.only(left: 12.0,),
                     border: InputBorder.none,
                     hintText: "Write a Task",
-                    hintStyle: TextStyle(color: Colors.black54),
+                    hintStyle: TextStyle(
+                        color: Colors.black54, fontSize: 16 * unitValueHeight),
                   ),
                 ),
               ),
@@ -102,7 +105,7 @@ class _AddTaskState extends State<AddTask> {
       String title = controller.text;
       controller.clear();
       textfieldFocus.unfocus();
-      await widget.taskbloc.addTask(title, widget.length, false);
+      await widget.taskbloc.addTask(title);
       //widget.taskbloc.updateTasks();
     }
   }
