@@ -3,7 +3,6 @@ from flask import request
 from Models import db, Task, User, Group
 
 
-
 class GroupMembers(Resource):
 
     # List Members of a Group
@@ -12,7 +11,7 @@ class GroupMembers(Resource):
         header = request.headers["Authorization"]
 
         if not header:
-            return {"Messege": "No group key!"}, 401
+            return {"status": "No group key!"}, 401
         else:
             group = Group.query.filter_by(group_key=header).first()
             if group:
@@ -27,9 +26,9 @@ class GroupMembers(Resource):
         json_data = request.get_json(force=True)
 
         if not json_data:
-            return {'Message': 'No input data provided'}, 400
+            return {'status': 'No input data provided'}, 400
         if not header:
-            return {"Messege": "No group key!"}, 401
+            return {"status": "No group key!"}, 401
         else:
             group = Group.query.filter_by(group_key=header).first()
             if group:
@@ -51,7 +50,7 @@ class GroupMembers(Resource):
                 else:
                     return {"status": 'Group is not public'}, 403
             else:
-                return {"Messege": "No Group Found with that group key"}, 404
+                return {"status": "No Group Found with that group key"}, 404
 
     """ Update Group Members: to be create if roles are implemented
     def put(self):
@@ -59,7 +58,7 @@ class GroupMembers(Resource):
         json_data = request.get_json(force=True)
 
         if not header:
-            return {"Messege": "No group key!"}, 400
+            return {"status": "No group key!"}, 400
         else:
             group = Group.query.filter_by(group_key=header).first()
             if group:
@@ -72,7 +71,7 @@ class GroupMembers(Resource):
                 result = Group.serialize(group)
                 return {"status": 'success', 'data': result}, 201
             else:
-                return {"Messege": "No Group with that group key"}, 402 
+                return {"status": "No Group with that group key"}, 402 
     """
 
     # Delete Members: Also, delete Member's tasks related to the specified Group
@@ -81,9 +80,9 @@ class GroupMembers(Resource):
         username = request.args.get('username')
 
         if not header:
-            return {"Messege": "No group key!"}, 401
+            return {"status": "No group key!"}, 401
         if not username:
-            return {'Message': 'No username provided'}, 400
+            return {'status': 'No username provided'}, 400
         else:
             group = Group.query.filter_by(group_key=header).first()
             if group:
@@ -91,9 +90,6 @@ class GroupMembers(Resource):
                 for m in group.members:
                     if m.username == username:
                         result = User.serialize(m)
-                        group_tasks = Task.query.filter_by(group_id=group.id)
-                        for task in group_tasks:
-                            db.session.delete(task)
                         group.members.remove(m)
                         if group.is_empty():
                             db.session.delete(group)
