@@ -3,6 +3,7 @@ from flask import request
 from Models import db, SubTask, Task, User
 import random
 import string
+from datetime import datetime
 
 
 class SubTasks(Resource):
@@ -12,7 +13,7 @@ class SubTasks(Resource):
         json_data = request.get_json(force=True)
 
         if not header:
-            return {"Messege": "No task key!"}, 400
+            return {"status": "No task key!"}, 400
         else:
             task = Task.query.filter_by(task_key=header).first()
             if task:
@@ -26,13 +27,7 @@ class SubTasks(Resource):
                 subtask = SubTask(
                     title=json_data['title'],
                     task_id=task.id,
-                    note=json_data['note'],
-                    completed=json_data['completed'],
-                    repeats=json_data['repeats'],
-                    group=json_data['group'],
-                    reminders=json_data['reminders'],
                     subtask_key=subtask_key,
-                    index=json_data['index'],
                 )
                 db.session.add(subtask)
                 db.session.commit()
@@ -40,7 +35,7 @@ class SubTasks(Resource):
                 result = SubTask.serialize(subtask)
                 return {"status": 'success', 'data': result}, 201
             else:
-                return {"Messege": "No task found with that key"}, 404
+                return {"status": "No task found with that key"}, 404
 
     # List Subtasks
     def get(self):
@@ -48,7 +43,7 @@ class SubTasks(Resource):
         header = request.headers["Authorization"]
 
         if not header:
-            return {"Messege": "No Task Key!"}, 401
+            return {"status": "No Task Key!"}, 401
         else:
             task = Task.query.filter_by(task_key=header).first()
             if task:
@@ -57,36 +52,31 @@ class SubTasks(Resource):
                     result.append(SubTask.serialize(subtask))
                 return {"status": 'success', 'data': result}, 200
             else:
-                return {"Message": "No task found with that task key"}, 404
+                return {"status": "No task found with that task key"}, 404
+
+        return "<h1>Subtasks!!</h1>"
 
     def put(self):
         header = request.headers["Authorization"]
         json_data = request.get_json(force=True)
         if not header:
-            return {"Messege": "No subtask key!"}, 401
+            return {"status": "No subtask key!"}, 401
         else:
             subtask = SubTask.query.filter_by(subtask_key=header).first()
             if subtask:
-                if (subtask.title != json_data['title']):
-                    subtask.title = json_data['title']
                 if (subtask.note != json_data['note']):
                     subtask.note = json_data['note']
                 if (subtask.completed != json_data['completed']):
                     subtask.completed = json_data['completed']
-                if (subtask.repeats != json_data['repeats']):
-                    subtask.repeats = json_data['repeats'],
-                if (subtask.group != json_data['group']):
-                    subtask.group = json_data['group']
-                if (subtask.reminders != json_data['reminders']):
-                    subtask.reminders = json_data['reminders']
-                if (subtask.index != json_data['index']):
-                    subtask.index = json_data['index']
+                if (subtask.due_date != datetime.fromisoformat(json_data['due_date'])):
+                    subtask.due_date = datetime.fromisoformat(
+                        json_data['due_date'])
                 db.session.commit()
                 result = SubTask.serialize(subtask)
                 return {"status": 'success', 'data': result}, 200
             else:
                 return {
-                    "Messege": "No SubTask Found with that subtask key"
+                    "status": "No SubTask Found with that subtask key"
                 }, 404
 
     # Delete Subtask
@@ -94,7 +84,7 @@ class SubTasks(Resource):
         header = request.headers["Authorization"]
 
         if not header:
-            return {"Messege": "No subtask key!"}, 401
+            return {"status": "No subtask key!"}, 401
         else:
             subtask = SubTask.query.filter_by(subtask_key=header).first()
             if subtask:
@@ -104,7 +94,7 @@ class SubTasks(Resource):
                 return {"status": 'success', 'data': result}, 200
             else:
                 return {
-                    "Message": 'No Subtask found with that subtask key'
+                    "status": 'No Subtask found with that subtask key'
                 }, 404
 
     def generate_key(self):
