@@ -6,50 +6,49 @@ import string
 
 
 class Users(Resource):
-    # Get User List
+    """ # Get User List
     def get(self):
         users = User.query.all()
         user_list = []
         for user in users:
             user_list.append(user.serialize())
-        return {"status": user_list}, 200
+        return {"status": user_list}, 200 """
 
     # Create New User (Sign Up/register User)
     def post(self):
         json_data = request.get_json(force=True)
 
         if not json_data:
-            return {'Message': 'No input data provided'}, 400
+            return {'status': 'No input data provided'}, 400
 
         user = User.query.filter_by(username=json_data['username']).first()
         if user:
-            return {'Message': 'Username is already taken'}, 409
+            return {'status': 'Username is already taken'}, 409
 
         user = User.query.filter_by(
             emailaddress=json_data['emailaddress']).first()
         if user:
-            return {'Message': 'Email address already exists'}, 409
+            return {'status': 'Email address already exists'}, 409
 
         user = User.query.filter_by(
             phonenumber=json_data['phonenumber']).first()
         if user:
-            return {'Message': 'Phone Number already exists'}, 409
+            return {'status': 'Phone Number already exists'}, 409
 
         api_key = self.generate_key()
-
         user = User.query.filter_by(api_key=api_key).first()
-        if user:
-            return {'Message': 'API key already exists'}, 409
+        while user:
+            api_key = self.generate_key()
+            user = User.query.filter_by(api_key=api_key).first()
 
         user = User(
             api_key=api_key,
             firstname=json_data['firstname'],
             lastname=json_data['lastname'],
-            phonenumber=json_data['phonenumber'],
-            avatar=json_data['avatar'],
-            emailaddress=json_data['emailaddress'],
-            password=json_data['password'],
             username=json_data['username'],
+            password=json_data['password'],
+            emailaddress=json_data['emailaddress'],
+            phonenumber=json_data['phonenumber'],
         )
         db.session.add(user)
         db.session.commit()
@@ -63,12 +62,12 @@ class Users(Resource):
         header = request.headers["Authorization"]
         json_data = request.get_json(force=True)
         if not header:
-            return {'Messege': "No API key!"}, 401
+            return {'status': "No API key!"}, 401
         else:
             user = User.query.filter_by(api_key=header).first()
             if user:
-                if user.password != json_data["currentPassword"]:
-                    return {"Message": "Incorrect Current Password"}
+                if user.verify_password(json_data["currentPassword"]):
+                    return {"status": "Incorrect Current Password"}
                 else:
                     if (user.username != json_data['username']):
                         user.username = json_data['username']
@@ -91,7 +90,7 @@ class Users(Resource):
                 return {"status": 'success', 'data': result}, 200
 
             else:
-                return {'Messege': "No User found with that api key"}, 404
+                return {'status': "No User found with that api key"}, 404
 
     # Generate new api key
     def generate_key(self):
