@@ -1,44 +1,37 @@
 import 'package:flutter/material.dart';
 import 'package:todolist/UI/pages/sidebar_pages/create_new_group_page.dart';
 import 'package:todolist/UI/tabs/list_groups_tab.dart';
-import 'package:todolist/bloc/blocs/user_bloc_provider.dart';
 
 import 'package:todolist/models/global.dart';
 import 'package:todolist/widgets/sidebar_widgets/sidebar_menu.dart';
 
 class HomePage extends StatefulWidget {
-  final String title;
-
-  HomePage({this.title});
-
+  static const routeName = '/home';
   @override
   _HomePageState createState() => _HomePageState();
 }
 
 class _HomePageState extends State<HomePage> {
   bool isMenuOpen = false;
-  @override
-  void initState() {
-    groupBloc.updateGroups();
-    super.initState();
-  }
+  late double unitHeightValue;
 
   @override
   Widget build(BuildContext context) {
+    unitHeightValue = MediaQuery.of(context).size.height * 0.001;
     return SafeArea(
       child: Scaffold(
         appBar: AppBar(
-          title: Text("TODO List", style: appTitleStyle),
+          title: Text("TODO List", style: appTitleStyle(unitHeightValue)),
           centerTitle: true,
           backgroundColor: Colors.white,
           elevation: 0,
-          iconTheme: IconThemeData(color: Colors.black),
           leading: IconButton(
             icon: isMenuOpen
-                ? Icon(Icons.arrow_back, color: darkBlueGradient, size: 32.0)
+                ? Icon(Icons.arrow_back,
+                    color: darkBlueGradient, size: 32.0 * unitHeightValue)
                 : Icon(
                     Icons.settings,
-                    size: 32.0,
+                    size: 32.0 * unitHeightValue,
                     color: darkBlueGradient,
                   ),
             onPressed: () {
@@ -52,16 +45,14 @@ class _HomePageState extends State<HomePage> {
               icon: Icon(
                 Icons.group_add,
                 color: darkBlueGradient,
-                size: 32.0,
+                size: 32 * unitHeightValue,
               ),
               onPressed: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                    builder: (context) => CreateGroupPage(),
-                  ),
-                );
+                Navigator.pushNamed(context, CreateGroupPage.routeName);
               }, //will go to Create a group Page
+            ),
+            SizedBox(
+              width: 10,
             )
           ],
         ),
