@@ -10,6 +10,8 @@ import 'package:todolist/widgets/global_widgets/background_color_container.dart'
 import 'package:todolist/widgets/global_widgets/custom_appbar.dart';
 
 class CreateGroupPage extends StatefulWidget {
+  static const routeName = '/create_group';
+
   @override
   _CreateGroupPageState createState() => _CreateGroupPageState();
 }
@@ -19,6 +21,7 @@ class _CreateGroupPageState extends State<CreateGroupPage> {
   int membersLength = 0;
   bool isPrivate = true;
   TextEditingController groupName = new TextEditingController();
+  late double unitHeightValue, unitWidthValue;
 
   @override
   void initState() {
@@ -31,32 +34,44 @@ class _CreateGroupPageState extends State<CreateGroupPage> {
 
   @override
   Widget build(BuildContext context) {
+    unitHeightValue = MediaQuery.of(context).size.height * 0.001;
+    unitWidthValue = MediaQuery.of(context).size.width * 0.001;
+
     return SafeArea(
       child: BackgroundColorContainer(
         startColor: lightBlue,
         endColor: lightBlueGradient,
-        widget: Scaffold(
-          appBar: CustomAppBar(
-            "New Group/Project",
-            actions: <Widget>[
-              TextButton(
-                onPressed: saveGroup,
-                child: Text(
-                  "Save",
-                  style: TextStyle(
-                      color: Colors.lightBlue,
-                      fontFamily: "Segoe UI",
-                      fontSize: 20,
-                      fontWeight: FontWeight.bold),
+        widget: GestureDetector(
+          onTap: () {
+            FocusScopeNode currentFocus = FocusScope.of(context);
+
+            if (!currentFocus.hasPrimaryFocus) {
+              currentFocus.unfocus();
+            }
+          },
+          child: Scaffold(
+            appBar: CustomAppBar(
+              "New Group/Project",
+              actions: <Widget>[
+                TextButton(
+                  onPressed: saveGroup,
+                  child: Text(
+                    "Save",
+                    style: TextStyle(
+                        color: Colors.lightBlue,
+                        fontFamily: "Segoe UI",
+                        fontSize: 20 * unitHeightValue,
+                        fontWeight: FontWeight.bold),
+                  ),
                 ),
-              ),
-            ],
-            fontSize: 24,
-          ),
-          backgroundColor: Colors.transparent,
-          body: Padding(
-            padding: const EdgeInsets.only(top: 15.0),
-            child: _buildStack(),
+              ],
+              fontSize: 24 * unitHeightValue,
+            ),
+            backgroundColor: Colors.transparent,
+            body: Padding(
+              padding: const EdgeInsets.only(top: 15.0),
+              child: _buildStack(),
+            ),
           ),
         ),
       ),
@@ -64,14 +79,15 @@ class _CreateGroupPageState extends State<CreateGroupPage> {
   }
 
   void saveGroup() async {
-    String groupKey = await repository.addGroup(groupName.text, !isPrivate);
+    String groupKey = await groupBloc.addGroup(groupName.text, !isPrivate);
     for (GroupMember member in newGroup.members) {
       try {
         await repository.addGroupMember(groupKey, member.username);
       } catch (e) {
-        print(e.message);
+        print(e);
       }
     }
+    await groupBloc.updateGroups();
     Navigator.pop(context);
   }
 
@@ -100,10 +116,10 @@ class _CreateGroupPageState extends State<CreateGroupPage> {
 
   CircleAvatar _buildAvatar() {
     return CircleAvatar(
-      radius: 50.0,
+      radius: 50.0 * unitHeightValue,
       child: Icon(
         Icons.group,
-        size: 62.0,
+        size: 62.0 * unitHeightValue,
       ),
     );
   }
@@ -111,7 +127,7 @@ class _CreateGroupPageState extends State<CreateGroupPage> {
   Container _buildGroupNameContainer() {
     return Container(
       //margin: const EdgeInsets.only(left: 100.0, right: 45.0, bottom: 20.0),
-      width: 250,
+      width: 250 * unitWidthValue,
       padding: EdgeInsets.only(left: 20),
       alignment: Alignment.center,
       decoration: BoxDecoration(
@@ -132,20 +148,20 @@ class _CreateGroupPageState extends State<CreateGroupPage> {
           fontFamily: 'Segoe UI',
           fontWeight: FontWeight.bold,
           color: lightBlue,
-          fontSize: 30,
+          fontSize: 30 * unitHeightValue,
         ),
         suffixIcon: Icon(
           Icons.edit,
           color: lightBlue,
+          size: 24 * unitHeightValue,
         ),
         isDense: true,
       ),
       style: TextStyle(
-        fontFamily: 'Segoe UI',
-        fontWeight: FontWeight.bold,
-        color: lightBlue,
-        fontSize: 30,
-      ),
+          fontFamily: 'Segoe UI',
+          fontWeight: FontWeight.bold,
+          color: lightBlue,
+          fontSize: 30 * unitHeightValue),
       onChanged: (groupName) => newGroup.name = groupName,
     );
   }
@@ -177,32 +193,35 @@ class _CreateGroupPageState extends State<CreateGroupPage> {
       Text(
         "MEMBERS",
         style: TextStyle(
-            fontFamily: 'Segoe UI',
-            fontWeight: FontWeight.bold,
-            color: darkBlue,
-            fontSize: 22),
+          fontFamily: 'Segoe UI',
+          fontWeight: FontWeight.bold,
+          color: darkBlue,
+          fontSize: 22 * unitHeightValue,
+        ),
       ),
-      SizedBox(width: 15),
+      SizedBox(width: 15 * unitWidthValue),
       CircleAvatar(
-        radius: 16,
+        radius: 16 * unitHeightValue,
         backgroundColor: darkBlue,
         child: Text(
           "${newGroup.members.length}",
           style: TextStyle(
-              color: Colors.white,
-              fontFamily: 'Segoe UI',
-              fontWeight: FontWeight.bold,
-              fontSize: 16),
+            color: Colors.white,
+            fontFamily: 'Segoe UI',
+            fontWeight: FontWeight.bold,
+            fontSize: 16 * unitHeightValue,
+          ),
         ),
       ),
       Spacer(),
       Text(
         "Personal",
         style: TextStyle(
-            fontFamily: 'Segoe UI',
-            fontWeight: FontWeight.bold,
-            color: Colors.black54,
-            fontSize: 20),
+          fontFamily: 'Segoe UI',
+          fontWeight: FontWeight.bold,
+          color: Colors.black54,
+          fontSize: 20 * unitHeightValue,
+        ),
       ),
       Switch(
           value: isPrivate,
@@ -221,16 +240,18 @@ class _CreateGroupPageState extends State<CreateGroupPage> {
       });
     });
     return Padding(
-      padding: EdgeInsets.only(top: 44.0, right: 24.0),
+      padding: EdgeInsets.only(
+          top: 44.0 * unitHeightValue, right: 24.0 * unitWidthValue),
       child: GridView.builder(
         gridDelegate: SliverGridDelegateWithMaxCrossAxisExtent(
-            maxCrossAxisExtent: 110,
+            maxCrossAxisExtent: 110 * unitWidthValue,
             childAspectRatio: 0.75,
-            crossAxisSpacing: 10,
-            mainAxisSpacing: 10.0),
+            crossAxisSpacing: 10 * unitWidthValue,
+            mainAxisSpacing: 10.0 * unitHeightValue),
         itemBuilder: (context, index) => Column(
           children: [
-            newGroup.members[index].cAvatar(radius: 34),
+            newGroup.members[index]
+                .cAvatar(radius: 34, unitHeightValue: unitHeightValue),
             Text(
               newGroup.members[index].firstname,
               overflow: TextOverflow.ellipsis,
@@ -263,7 +284,7 @@ class _CreateGroupPageState extends State<CreateGroupPage> {
                   ),
                 );
               },
-              child: Icon(Icons.arrow_forward, size: 36),
+              child: Icon(Icons.arrow_forward, size: 36 * unitHeightValue),
             ),
           );
   }
