@@ -1,9 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:todolist/models/global.dart';
 
+// ignore: must_be_immutable
 class HowToPage extends StatelessWidget {
+  late double unitHeightValue;
   @override
   Widget build(BuildContext context) {
+    unitHeightValue = MediaQuery.of(context).size.height * 0.001;
     return Scaffold(
       backgroundColor: lightBlueGradient,
       appBar: AppBar(
@@ -28,15 +31,16 @@ class HowToPage extends StatelessWidget {
         child: RichText(
           text: TextSpan(
             text: 'How to Use:',
-            style: cardTitleStyle,
+            style: cardTitleStyle(unitHeightValue),
             children: <TextSpan>[
               TextSpan(
                 text: '\nAdd: ',
                 style: TextStyle(
-                    fontFamily: 'Avenir',
-                    fontWeight: FontWeight.bold,
-                    color: darkBlueGradient,
-                    fontSize: 26),
+                  fontFamily: 'Avenir',
+                  fontWeight: FontWeight.bold,
+                  color: darkBlueGradient,
+                  fontSize: 26 * unitHeightValue,
+                ),
               ),
               TextSpan(
                 text: '\nClick the large + sign to add a task',
@@ -44,7 +48,7 @@ class HowToPage extends StatelessWidget {
                     fontFamily: 'Avenir',
                     fontWeight: FontWeight.bold,
                     color: darkBlue,
-                    fontSize: 18,
+                    fontSize: 18 * unitHeightValue,
                     height: 1.8),
               ),
               TextSpan(
@@ -53,7 +57,7 @@ class HowToPage extends StatelessWidget {
                     fontFamily: 'Avenir',
                     fontWeight: FontWeight.bold,
                     color: darkBlueGradient,
-                    fontSize: 26),
+                    fontSize: 26 * unitHeightValue),
               ),
               TextSpan(
                 text: '\nSwipe Left on a List Item.',
@@ -61,7 +65,7 @@ class HowToPage extends StatelessWidget {
                     fontFamily: 'Avenir',
                     fontWeight: FontWeight.bold,
                     color: darkBlue,
-                    fontSize: 18,
+                    fontSize: 18 * unitHeightValue,
                     height: 1.8),
               ),
               TextSpan(
@@ -70,7 +74,7 @@ class HowToPage extends StatelessWidget {
                     fontFamily: 'Avenir',
                     fontWeight: FontWeight.bold,
                     color: darkBlueGradient,
-                    fontSize: 26),
+                    fontSize: 26 * unitHeightValue),
               ),
               TextSpan(
                 text:
@@ -79,7 +83,7 @@ class HowToPage extends StatelessWidget {
                     fontFamily: 'Avenir',
                     fontWeight: FontWeight.bold,
                     color: darkBlue,
-                    fontSize: 18,
+                    fontSize: 18 * unitHeightValue,
                     height: 1.8),
               ),
               TextSpan(
@@ -88,7 +92,7 @@ class HowToPage extends StatelessWidget {
                     fontFamily: 'Avenir',
                     fontWeight: FontWeight.bold,
                     color: darkBlueGradient,
-                    fontSize: 26),
+                    fontSize: 26 * unitHeightValue),
               ),
               TextSpan(
                 text:
@@ -97,16 +101,17 @@ class HowToPage extends StatelessWidget {
                     fontFamily: 'Avenir',
                     fontWeight: FontWeight.bold,
                     color: darkBlue,
-                    fontSize: 18,
+                    fontSize: 18 * unitHeightValue,
                     height: 1.8),
               ),
               TextSpan(
                 text: '\nSubtasks: ',
                 style: TextStyle(
-                    fontFamily: 'Avenir',
-                    fontWeight: FontWeight.bold,
-                    color: darkBlueGradient,
-                    fontSize: 26),
+                  fontFamily: 'Avenir',
+                  fontWeight: FontWeight.bold,
+                  color: darkBlueGradient,
+                  fontSize: 26 * unitHeightValue,
+                ),
               ),
               TextSpan(
                 text: '\nClick a task to view it\'s subtasks',
@@ -114,7 +119,7 @@ class HowToPage extends StatelessWidget {
                     fontFamily: 'Avenir',
                     fontWeight: FontWeight.bold,
                     color: darkBlue,
-                    fontSize: 18,
+                    fontSize: 18 * unitHeightValue,
                     height: 1.8),
               ),
             ],
