@@ -1,6 +1,3 @@
-//import 'package:todolist/models/tasks.dart';
-
-import 'package:flutter/cupertino.dart';
 import 'package:todolist/models/group.dart';
 import 'package:todolist/models/groupmember.dart';
 import 'package:todolist/models/subtasks.dart';
@@ -11,8 +8,8 @@ import 'package:todolist/models/user.dart';
 import 'package:todolist/models/tasks.dart';
 
 class UserBloc {
-  final _userGetter = PublishSubject<User>();
-  User _user;
+  final PublishSubject<User> _userGetter = PublishSubject<User>();
+  User _user = new User.blank();
 
   UserBloc._privateConstructor();
 
@@ -22,29 +19,34 @@ class UserBloc {
     return _instance;
   }
 
-  PublishSubject<User> get getUser => _userGetter.stream;
+  Stream<User> get getUser => _userGetter.stream;
 
   User getUserObject() {
     return _user;
   }
 
-  registerUser(String username, String password, String email, String firstname,
-      String lastname, String phonenumber, ImageProvider avatar) async {
-    _user = await repository.registerUser(
-        username, password, email, firstname, lastname, phonenumber, avatar);
-    _userGetter.sink.add(_user);
+  Future<void> registerUser(String username, String password, String email,
+      String firstname, String lastname, String phonenumber, avatar) async {
+    try {
+      _user = await repository.registerUser(
+          username, password, email, firstname, lastname, phonenumber, avatar);
+      _userGetter.sink.add(_user);
+    } catch (e) {
+      throw e;
+    }
   }
 
-  signinUser(String username, String password, String apiKey) async {
+  Future<void> signinUser(
+      String username, String password, String apiKey) async {
     try {
       _user = await repository.signinUser(username, password, apiKey);
+      _userGetter.sink.add(_user);
     } catch (e) {
-      throw Exception(e.message);
+      throw e;
     }
-    _userGetter.sink.add(_user);
   }
 
-  updateUserProfile(
+  Future<void> updateUserProfile(
       String currentPassword,
       String newPassword,
       String email,
@@ -52,12 +54,13 @@ class UserBloc {
       String firstname,
       String lastname,
       String phonenumber,
-      ImageProvider avatar) async {
+      avatar) async {
     try {
       _user = await repository.updateUserProfile(currentPassword, newPassword,
           email, username, firstname, lastname, phonenumber, avatar);
+      _userGetter.sink.add(_user);
     } catch (e) {
-      throw Exception(e.message);
+      throw e;
     }
   }
 
@@ -79,11 +82,6 @@ class GroupBloc {
   factory GroupBloc() {
     return _instance;
   }
-  /* Testing GroupBloc 
-  Future<List<Group>> testGroupList() async{
-    return await repository.getUserGroups(_apiKey);
-  } 
-  */
 
   Stream<List<Group>> get getGroups {
     updateGroups();
@@ -94,12 +92,19 @@ class GroupBloc {
     return _groups;
   }
 
-  Future<Null> deleteGroup(String groupKey) async {
+  Future<void> deleteGroup(String groupKey) async {
     await repository.deleteGroup(groupKey);
     await updateGroups();
   }
 
-  Future<Null> updateGroups() async {
+  Future<String> addGroup(String groupName, bool isPublic) async {
+    String groupKey = await repository.addGroup(groupName, isPublic);
+    await updateGroups();
+    return groupKey;
+  }
+
+  Future<void> updateGroups() async {
+    await Future<void>.delayed(const Duration(milliseconds: 150));
     _groups = await repository.getUserGroups();
     _groupSubject.add(_groups);
   }
@@ -110,16 +115,14 @@ class GroupMemberBloc {
   String groupKey;
   var _groupMembers = <GroupMember>[];
 
-  GroupMemberBloc(String groupKey) {
-    this.groupKey = groupKey;
-    _updateGroupMembers(groupKey).then((_) {
-      _groupMemberSubject.add(_groupMembers);
-    });
+  GroupMemberBloc(String groupKey) : this.groupKey = groupKey {
+    _updateGroupMembers(groupKey);
   }
 
   Stream<List<GroupMember>> get getGroupMembers => _groupMemberSubject.stream;
-  Future<Null> _updateGroupMembers(String groupKey) async {
+  Future<void> _updateGroupMembers(String groupKey) async {
     _groupMembers = await repository.getGroupMembers(groupKey);
+    _groupMemberSubject.add(_groupMembers);
   }
 }
 
@@ -127,58 +130,69 @@ class TaskBloc {
   final _taskSubject = BehaviorSubject<List<Task>>();
   String _groupKey;
 
-  TaskBloc(String groupKey) {
-    this._groupKey = groupKey;
+  TaskBloc(String groupKey) : this._groupKey = groupKey {
     updateTasks();
   }
 
   Stream<List<Task>> get getTasks => _taskSubject.stream;
 
-  Future<Null> addTask(String taskName, int index, bool completed) async {
-    await repository.addTask(taskName, this._groupKey, index, completed);
-    await Future<void>.delayed(const Duration(milliseconds: 300));
+  Future<void> addTask(String taskName) async {
+    await repository.addTask(taskName, this._groupKey);
     await updateTasks();
   }
 
-  Future<Null> deleteTask(String taskKey) async {
+  Future<void> deleteTask(String taskKey) async {
     await repository.deleteTask(taskKey);
-    await Future<void>.delayed(const Duration(milliseconds: 400));
     await updateTasks();
   }
 
-  Future<Null> updateTasks() async {
-    await repository.getTasks(this._groupKey).then((tasks) {
-      print("Updated List" + tasks.toString());
-      _taskSubject.add(tasks);
-    });
+  Future<void> updateTasks() async {
+    await Future<void>.delayed(const Duration(milliseconds: 400));
+    List<Task> tasks = await repository.getTasks(this._groupKey);
+    _taskSubject.add(tasks);
   }
 }
 
 class SubtaskBloc {
   final _subtaskSubject = BehaviorSubject<List<Subtask>>();
-  String _taskKey;
+  Task _task;
 
-  SubtaskBloc(String taskKey) {
-    this._taskKey = taskKey;
+  SubtaskBloc(Task task) : this._task = task {
     _updateSubtasks();
   }
 
   Stream<List<Subtask>> get getSubtasks => _subtaskSubject.stream;
 
-  Future<Null> addSubtask(String subtaskName, int index, bool completed) async {
-    await repository.addSubtask(_taskKey, subtaskName, index, completed);
-    await Future<void>.delayed(const Duration(milliseconds: 300));
-    await _updateSubtasks();
+  Future<void> addSubtask(String subtaskName) async {
+    Future.wait(
+      [
+        repository.addSubtask(_task.taskKey, subtaskName),
+        _updateSubtasks(wait: true),
+      ],
+    );
   }
 
-  Future<Null> deleteSubtask(String subtaskKey) async {
+  Future<void> deleteSubtask(String subtaskKey) async {
     await repository.deleteSubtask(subtaskKey);
-    await Future<void>.delayed(const Duration(milliseconds: 300));
-    await _updateSubtasks();
+    await _updateSubtasks(wait: true);
   }
 
-  Future<Null> _updateSubtasks() async {
-    await repository.getSubtasks(_taskKey).then((subtasks){_subtaskSubject.add(subtasks);});
+  Future<void> updateSubtaskInfo(Subtask subtask) async {
+    Future.wait(
+      [
+        repository.updateSubtask(subtask),
+        _updateSubtasks(),
+      ],
+    );
+    //await repository.updateSubtask(subtask);
+    //await _updateSubtasks();
+  }
+
+  Future<void> _updateSubtasks({bool wait = false}) async {
+    if (wait) await Future<void>.delayed(const Duration(milliseconds: 400));
+
+    List<Subtask> subtasks = await repository.getSubtasks(_task);
+    _subtaskSubject.add(subtasks);
   }
 }
 
