@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'package:flutter/cupertino.dart';
 import 'package:http/http.dart' show Client;
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:todolist/models/group.dart';
@@ -16,39 +15,37 @@ class ApiProvider {
   //static String baseURL = "http://10.0.2.2:5000/api";
   static String stageHost = 'taskmanager-group-stage.herokuapp.com';
   static String productionHost = 'taskmanager-group-pro.herokuapp.com';
-  static String localhost = "http://10.0.2.2:5000/api";
+  static String localhost = "10.0.2.2:5000";
   Uri signinURL = Uri(scheme: 'https', host: stageHost, path: '/api/signin');
   Uri userURL = Uri(scheme: 'https', host: stageHost, path: '/api/user');
   Uri taskURL = Uri(scheme: 'https', host: stageHost, path: '/api/tasks');
   Uri subtaskURL = Uri(scheme: 'https', host: stageHost, path: '/api/subtasks');
   Uri groupURL = Uri(scheme: 'https', host: stageHost, path: '/api/group');
   Uri groupmemberURL =
-      Uri(scheme: 'https', host: stageHost, path: '/api/groupmember');
-  Uri searchURL = Uri(scheme: 'https', host: stageHost, path: '/api/search');
+      Uri(scheme: 'http', host: stageHost, path: '/api/groupmember');
+  Uri searchURL = Uri(scheme: 'http', host: stageHost, path: '/api/search');
+  Uri assignedtouserhURL =
+      Uri(scheme: 'http', host: stageHost, path: '/api/assignedtouser');
 
-  String apiKey;
+  String apiKey = '';
 
   // User CRUD Functions
   /// Sign Up
-  Future<User> registerUser(
-      String username,
-      String password,
-      String email,
-      String firstname,
-      String lastname,
-      String phonenumber,
-      ImageProvider avatar) async {
-    final response = await client.post(userURL,
-        // headers: "",
-        body: jsonEncode({
-          "emailaddress": email,
-          "username": username,
-          "password": password,
-          "firstname": firstname,
-          "lastname": lastname,
-          "phonenumber": phonenumber,
-          "avatar": avatar,
-        }));
+  Future<User> registerUser(String username, String password, String email,
+      String firstname, String lastname, String phonenumber, avatar) async {
+    final response = await client.post(
+      userURL,
+      // headers: "",
+      body: jsonEncode({
+        "emailaddress": email,
+        "username": username,
+        "password": password,
+        "firstname": firstname,
+        "lastname": lastname,
+        "phonenumber": phonenumber,
+        "avatar": avatar,
+      }),
+    );
     final Map result = json.decode(response.body);
     if (response.statusCode == 201) {
       // If the call to the server was successful, parse the JSON
@@ -56,26 +53,29 @@ class ApiProvider {
       return User.fromJson(result["data"]);
     } else {
       // If that call was not successful, throw an error.
-      throw Exception(result["Message"]);
+      throw Exception(result["status"]);
     }
   }
 
   /// Sign User In using username and password or API_Key
   Future signinUser(String username, String password, String apiKey) async {
-    final response = await client.post(signinURL,
-        headers: {"Authorization": apiKey},
-        body: jsonEncode({
-          "username": username,
-          "password": password,
-        }));
+    final response = await client.post(
+      signinURL,
+      headers: {"Authorization": apiKey},
+      body: jsonEncode({
+        "username": username,
+        "password": password,
+      }),
+    );
     final Map result = json.decode(response.body);
     if (response.statusCode == 200) {
       // If the call to the server was successful, parse the JSON
       await saveApiKey(result["data"]["api_key"]);
+      await Future<void>.delayed(const Duration(milliseconds: 200));
       return User.fromJson(result["data"]);
     } else {
       // If that call was not successful, throw an error.
-      throw Exception(result["Message"]);
+      throw Exception(result["status"]);
     }
   }
 
@@ -88,27 +88,29 @@ class ApiProvider {
       String firstname,
       String lastname,
       String phonenumber,
-      ImageProvider avatar) async {
-    final response = await client.put(userURL,
-        headers: {"Authorization": apiKey},
-        body: jsonEncode({
-          "currentPassword": currentPassword,
-          "newPassword": newPassword,
-          "email": email,
-          "username": username,
-          "firstname": firstname,
-          "lastname": lastname,
-          "phonenumber": phonenumber,
-          "avatar": avatar,
-        }));
+      avatar) async {
+    final response = await client.put(
+      userURL,
+      headers: {"Authorization": apiKey},
+      body: jsonEncode({
+        "currentPassword": currentPassword,
+        "newPassword": newPassword,
+        "email": email,
+        "username": username,
+        "firstname": firstname,
+        "lastname": lastname,
+        "phonenumber": phonenumber,
+        "avatar": avatar,
+      }),
+    );
     final Map result = json.decode(response.body);
     if (response.statusCode == 200) {
-      print("User Profile Updated");
+      //print("User Profile Updated");
       return User.fromJson(result["data"]);
     } else {
       // If that call was not successful, throw an error.
       print(json.decode(response.body));
-      throw Exception(result["Message"]);
+      throw Exception(result["status"]);
     }
   }
 
@@ -125,21 +127,23 @@ class ApiProvider {
       final Map result = json.decode(response.body);
       if (response.statusCode == 200) {
         // If the call to the server was successful, parse the JSON
-        for (Map json_ in result["data"]) {
+        for (Map<String, dynamic> json_ in result["data"]) {
           try {
             Group group = Group.fromJson(json_);
+            //print("Get User Groups: ${group.groupKey}");
             group.members = await getGroupMembers(group.groupKey);
-            group.tasks = await getTasks(group.groupKey);
+            //print("--------------End members-------------");
+            //group.tasks = await getTasks(group.groupKey);
+            //print("--------------End tasks-------------");
             groups.add(group);
           } catch (Exception) {
             print(Exception);
-            throw Exception;
           }
         }
         return groups;
       } else {
         // If that call was not successful, throw an error.
-        throw Exception(result["Message"]);
+        throw Exception(result["status"]);
       }
     }
     return groups;
@@ -148,22 +152,41 @@ class ApiProvider {
   /// Add a Group
   Future addGroup(String groupName, bool isPublic) async {
     print(groupURL.toString());
-    final response = await client.post(groupURL,
-        headers: {"Authorization": apiKey},
-        body: jsonEncode({
-          "name": groupName,
-          "is_public": isPublic,
-        }));
+    final response = await client.post(
+      groupURL,
+      headers: {"Authorization": apiKey},
+      body: jsonEncode({
+        "name": groupName,
+        "is_public": isPublic,
+      }),
+    );
     if (response.statusCode == 201) {
       final Map result = json.decode(response.body);
       Group addedGroup = Group.fromJson(result["data"]);
-      print("Group: " + addedGroup.name + " added");
+      //print("Group: " + addedGroup.name + " added");
       return addedGroup.groupKey;
     } else {
       // If that call was not successful, throw an error.
       final Map result = json.decode(response.body);
-      print(result["Message"]);
-      throw Exception(result["Message"]);
+      print(result["status"]);
+      throw Exception(result["status"]);
+    }
+  }
+
+  /// Update Group Info
+  Future<bool> updateGroup(Group group) async {
+    final response = await client.patch(
+      groupURL,
+      headers: {"Authorization": group.groupKey},
+      body: jsonEncode({"name": group.name, "is_public": group.isPublic}),
+    );
+    final Map result = json.decode(response.body);
+    if (response.statusCode == 200) {
+      return true;
+      //print("Task ${task.title} Updated");
+    } else {
+      // If that call was not successful, throw an error.
+      throw Exception(result["status"]);
     }
   }
 
@@ -175,17 +198,18 @@ class ApiProvider {
     );
     if (response.statusCode == 200) {
       // If the call to the server was successful
-      print("Group deleted");
+      //print("Group deleted");
     } else {
       // If that call was not successful, throw an error.
       final Map result = json.decode(response.body);
-      throw Exception(result["Message"]);
+      throw Exception(result["status"]);
     }
   }
 
 // GroupMember CRUD Functions
   /// Get a list of the Group's Members.
   Future<List<GroupMember>> getGroupMembers(String groupKey) async {
+    //print("$groupKey");
     final response = await client.get(
       groupmemberURL,
       headers: {"Authorization": groupKey},
@@ -194,7 +218,7 @@ class ApiProvider {
     if (response.statusCode == 200) {
       // If the call to the server was successful, parse the JSON
       List<GroupMember> groupMembers = [];
-      for (Map json_ in result["data"]) {
+      for (Map<String, dynamic> json_ in result["data"]) {
         try {
           groupMembers.add(GroupMember.fromJson(json_));
         } catch (Exception) {
@@ -202,10 +226,11 @@ class ApiProvider {
           //throw Exception;
         }
       }
+      //print("getGroupMembers: " +groupMembers.toString() +" @" +DateTime.now().toString());
       return groupMembers;
     } else {
       // If that call was not successful, throw an error.
-      throw Exception(result["Message"]);
+      throw Exception(result["status"]);
     }
   }
 
@@ -214,20 +239,21 @@ class ApiProvider {
   /// * Username: Group Member's Username to be added
   Future addGroupMember(String groupKey, String username) async {
     print(groupmemberURL.toString());
-    final response = await client.post(groupmemberURL,
-        headers: {"Authorization": groupKey},
-        body: jsonEncode({
-          "username": username,
-        }));
+    final response = await client.post(
+      groupmemberURL,
+      headers: {"Authorization": groupKey},
+      body: jsonEncode({
+        "username": username,
+      }),
+    );
     final Map result = json.decode(response.body);
-    print(result.toString());
     if (response.statusCode == 201) {
       GroupMember addedGroupMember = GroupMember.fromJson(result["data"]);
       print("User ${addedGroupMember.username} added to GroupKey: $groupKey");
     } else {
       // If that call was not successful, throw an error.
-      print(result["Message"]);
-      throw Exception(result["Message"]);
+      print(result["status"]);
+      throw Exception(result["status"]);
     }
   }
 
@@ -243,13 +269,13 @@ class ApiProvider {
     );
     if (response.statusCode == 200) {
       // If the call to the server was successful
-      print("Group Member $username deleted");
+      //print("Group Member $username deleted");
     } else if (response.statusCode == 400 ||
         response.statusCode == 401 ||
         response.statusCode == 404) {
       // If that call was not successful, throw an error.
       final Map result = json.decode(response.body);
-      throw Exception(result["Message"]);
+      throw Exception(result["status"]);
     }
   }
 
@@ -257,6 +283,7 @@ class ApiProvider {
   /// Get a list of the Group's Tasks
   /// * GroupKey: Unique group identifier
   Future<List<Task>> getTasks(String groupKey) async {
+    //print("$groupKey");
     final response = await client.get(
       taskURL,
       headers: {"Authorization": groupKey},
@@ -265,10 +292,10 @@ class ApiProvider {
     if (response.statusCode == 200) {
       // If the call to the server was successful, parse the JSON
       List<Task> tasks = [];
-      for (Map json_ in result["data"]) {
+      for (Map<String, dynamic> json_ in result["data"]) {
         try {
           Task task = Task.fromJson(json_);
-          task.subtasks = await getSubtasks(task.taskKey);
+          //task.subtasks = await getSubtasks(task.taskKey);
           tasks.add(task);
         } catch (Exception) {
           print(Exception);
@@ -278,7 +305,7 @@ class ApiProvider {
       return tasks;
     } else {
       // If that call was not successful, throw an error.
-      throw Exception(result["Message"]);
+      throw Exception(result["status"]);
     }
   }
 
@@ -288,40 +315,31 @@ class ApiProvider {
   /// * GroupKey: Unique Group Identifier
   /// * Index: Position within Group's task list
   /// * Completed: True or False, Has the task been completed
-  Future addTask(
-      String taskName, String groupKey, int index, bool completed) async {
-    final response = await client.post(taskURL,
-        headers: {"Authorization": apiKey},
-        body: jsonEncode({
-          "title": taskName,
-          "group_key": groupKey,
-          "index": index,
-          "completed": completed
-        }));
+  Future addTask(String taskName, String groupKey) async {
+    final response = await client.post(
+      taskURL,
+      headers: {"Authorization": apiKey},
+      body: jsonEncode({"title": taskName, "group_key": groupKey}),
+    );
     if (response.statusCode == 201) {
-      print("Task " + taskName + " added @" + DateTime.now().toString());
+      //print("Task " + taskName + " added @" + DateTime.now().toString());
     } else {
       // If that call was not successful, throw an error.
       final Map result = json.decode(response.body);
-      print(result["Message"]);
-      throw Exception(result["Message"]);
+      print(result["status"]);
+      throw Exception(result["status"]);
     }
   }
 
   /// Update Task Info
   Future updateTask(Task task) async {
-    final response = await client.put(taskURL,
-        headers: {"Authorization": task.taskKey},
-        body: jsonEncode({
-          "title": task.title,
-          "note": task.note,
-          "repeats": task.repeats,
-          "completed": task.completed,
-          "reminders": task.reminders,
-          "index": task.index
-        }));
+    final response = await client.put(
+      taskURL,
+      headers: {"Authorization": task.taskKey},
+      body: jsonEncode({"completed": task.completed}),
+    );
     if (response.statusCode == 200) {
-      print("Task ${task.title} Updated");
+      //print("Task ${task.title} Updated");
     } else {
       // If that call was not successful, throw an error.
       print(json.decode(response.body));
@@ -338,28 +356,34 @@ class ApiProvider {
     );
     if (response.statusCode == 200) {
       // If the call to the server was successful
-      print("Task deleted @" + DateTime.now().toString());
+      //print("Task deleted @" + DateTime.now().toString());
     } else {
       // If that call was not successful, throw an error.
       final Map result = json.decode(response.body);
-      throw Exception(result["Message"]);
+      throw Exception(result["status"]);
     }
   }
 
 //Subtask CRUD Functions
   //Get Subtasks
-  Future<List<Subtask>> getSubtasks(String taskKey) async {
+  Future<List<Subtask>> getSubtasks(Task task) async {
     final response = await client.get(
       subtaskURL,
-      headers: {"Authorization": taskKey},
+      headers: {"Authorization": task.taskKey},
     );
     final Map result = json.decode(response.body);
     if (response.statusCode == 200) {
       // If the call to the server was successful, parse the JSON
       List<Subtask> subtasks = [];
-      for (Map json_ in result["data"]) {
+      for (Map<String, dynamic> json_ in result["data"]) {
         try {
-          subtasks.add(Subtask.fromJson(json_));
+          Subtask subtask = Subtask.fromJson(json_);
+          subtask.deadline = json_['due_date'] == null
+              ? DateTime.now()
+              : DateTime.parse(json_['due_date']);
+          subtask.assignedTo =
+              await getUsersAssignedToSubtask(subtask.subtaskKey);
+          subtasks.add(subtask);
         } catch (Exception) {
           print(Exception);
         }
@@ -367,52 +391,47 @@ class ApiProvider {
       return subtasks;
     } else {
       // If that call was not successful, throw an error.
-      throw Exception(result["Message"]);
+      throw Exception(result["status"]);
     }
   }
 
   //Add Subtask
-  Future addSubtask(
-      String taskKey, String subtaskName, int index, bool completed) async {
-    final response = await client.post(subtaskURL,
-        headers: {"Authorization": taskKey},
-        body: jsonEncode({
-          "title": subtaskName,
-          "note": "",
-          "completed": completed,
-          "repeats": "",
-          "group": "",
-          "reminders": "",
-          "index": index
-        }));
+  Future addSubtask(String taskKey, String subtaskName) async {
+    final response = await client.post(
+      subtaskURL,
+      headers: {"Authorization": taskKey},
+      body: jsonEncode({
+        "title": subtaskName,
+      }),
+    );
     if (response.statusCode == 201) {
-      print("Subtask " + subtaskName + " added @" + DateTime.now().toString());
+      await Future<void>.delayed(const Duration(milliseconds: 400));
+      //print("Subtask " + subtaskName + " added @" + DateTime.now().toString());
     } else {
+      print(response.statusCode);
       // If that call was not successful, throw an error.
       final Map result = json.decode(response.body);
-      throw Exception(result["Message"]);
+      throw Exception(result["status"]);
     }
   }
 
   //Update Subtask
   Future updateSubtask(Subtask subtask) async {
-    final response = await client.put(subtaskURL,
-        headers: {"Authorization": subtask.subtaskKey},
-        body: jsonEncode({
-          "title": subtask.title,
-          "note": subtask.note,
-          "repeats": subtask.repeats,
-          "completed": subtask.completed,
-          "group": subtask.group,
-          "reminders": subtask.reminders,
-          "index": subtask.index
-        }));
+    final response = await client.put(
+      subtaskURL,
+      headers: {"Authorization": subtask.subtaskKey},
+      body: jsonEncode({
+        "note": subtask.note,
+        "completed": subtask.completed,
+        "due_date": subtask.deadline.toIso8601String()
+      }),
+    );
     if (response.statusCode == 200) {
       print("Subtask " + subtask.title + " Updated");
     } else {
       // If that call was not successful, throw an error.
       final Map result = json.decode(response.body);
-      throw Exception(result["Message"]);
+      throw Exception(result["status"]);
     }
   }
 
@@ -424,26 +443,28 @@ class ApiProvider {
     );
     if (response.statusCode == 200) {
       // If the call to the server was successful
-      print("Subtask deleted @" + DateTime.now().toString());
+      //print("Subtask deleted @" + DateTime.now().toString());
     } else {
       // If that call was not successful, throw an error.
       final Map result = json.decode(response.body);
-      throw Exception(result["Message"]);
+      throw Exception(result["status"]);
     }
   }
 
   //Search API Calls
   Future<List<GroupMember>> searchUser(String searchTerm) async {
-    final response = await client.post(searchURL,
-        headers: {"Authorization": apiKey},
-        body: jsonEncode({
-          "search_term": searchTerm,
-        }));
+    final response = await client.post(
+      searchURL,
+      headers: {"Authorization": apiKey},
+      body: jsonEncode({
+        "search_term": searchTerm,
+      }),
+    );
     final Map result = json.decode(response.body);
     if (response.statusCode == 200) {
       // If the call to the server was successful, parse the JSON
       List<GroupMember> searchResults = [];
-      for (Map json_ in result["data"]) {
+      for (Map<String, dynamic> json_ in result["data"]) {
         try {
           searchResults.add(GroupMember.fromJson(json_));
         } catch (Exception) {
@@ -454,7 +475,77 @@ class ApiProvider {
       return searchResults;
     } else {
       // If that call was not successful, throw an error.
-      throw Exception(result["Message"]);
+      throw Exception(result["status"]);
+    }
+  }
+
+  ///AssignedToUser API Calls
+  ///GET
+  Future<List<GroupMember>> getUsersAssignedToSubtask(String subtaskKey) async {
+    final response = await client.get(
+      assignedtouserhURL,
+      headers: {"Authorization": subtaskKey},
+    );
+    final Map result = json.decode(response.body);
+    if (response.statusCode == 200) {
+      // If the call to the server was successful, parse the JSON
+      List<GroupMember> groupMembers = [];
+      for (Map<String, dynamic> json_ in result["data"]) {
+        try {
+          groupMembers.add(GroupMember.fromJson(json_));
+        } catch (Exception) {
+          print(Exception);
+          //throw Exception;
+        }
+      }
+      //print("getGroupMembers: " +groupMembers.toString() +" @" +DateTime.now().toString());
+      return groupMembers;
+    } else {
+      // If that call was not successful, throw an error.
+      throw Exception(result["status"]);
+    }
+  }
+
+  /// Post: Assign a Group Member to Subtask
+  /// * SubtaskKey: Unique Subtask Identifier
+  /// * Username: Group Member's Username to be added
+  Future assignSubtaskToUser(String subtaskKey, String username) async {
+    Uri assignURLQuery =
+        assignedtouserhURL.replace(query: "username=$username");
+    final response = await client.post(
+      assignURLQuery,
+      headers: {"Authorization": subtaskKey},
+    );
+    final Map result = json.decode(response.body);
+    if (response.statusCode == 201) {
+      print(
+          "User ${GroupMember.fromJson(result["data"]).username} assigned to SubtaskKey: $subtaskKey");
+    } else {
+      // If that call was not successful, throw an error.
+      print(result["status"]);
+      throw Exception(result["status"]);
+    }
+  }
+
+  /// Delete: Unssign a Group Member to Subtask
+  /// * SubtaskKey: Unique Subtask Identifier
+  /// * Username: Group Member's Username to be added
+  Future unassignSubtaskToUser(String subtaskKey, String username) async {
+    Uri assignURLQuery =
+        assignedtouserhURL.replace(query: "username=$username");
+    final response = await client.delete(
+      assignURLQuery,
+      headers: {"Authorization": subtaskKey},
+    );
+    if (response.statusCode == 200) {
+      // If the call to the server was successful
+      //print("Group Member $username deleted");
+    } else if (response.statusCode == 400 ||
+        response.statusCode == 401 ||
+        response.statusCode == 404) {
+      // If that call was not successful, throw an error.
+      final Map result = json.decode(response.body);
+      throw Exception(result["status"]);
     }
   }
 
