@@ -1,6 +1,6 @@
 import 'dart:async';
-import 'package:flutter/cupertino.dart';
 import 'package:todolist/models/group.dart';
+import 'package:todolist/models/groupmember.dart';
 import 'package:todolist/models/subtasks.dart';
 import 'package:todolist/models/tasks.dart';
 
@@ -10,14 +10,8 @@ import 'package:todolist/models/user.dart';
 class Repository {
   final apiProvider = ApiProvider();
 
-  Future<User> registerUser(
-          String username,
-          String password,
-          String email,
-          String firstname,
-          String lastname,
-          String phonenumber,
-          ImageProvider avatar) =>
+  Future<User> registerUser(String username, String password, String email,
+          String firstname, String lastname, String phonenumber, avatar) =>
       apiProvider.registerUser(
           username, password, email, firstname, lastname, phonenumber, avatar);
 
@@ -32,7 +26,7 @@ class Repository {
           String firstname,
           String lastname,
           String phonenumber,
-          ImageProvider avatar) =>
+          avatar) =>
       apiProvider.updateUserProfile(currentPassword, newPassword, email,
           username, firstname, lastname, phonenumber, avatar);
 
@@ -45,6 +39,8 @@ class Repository {
 
   Future addGroup(String groupName, bool isPublic) =>
       apiProvider.addGroup(groupName, isPublic);
+
+   Future<bool> updateGroup(Group group) => apiProvider.updateGroup(group);
 
   Future<dynamic> deleteGroup(String groupKey) =>
       apiProvider.deleteGroup(groupKey);
@@ -62,9 +58,8 @@ class Repository {
   //Tasks
   Future getTasks(String groupKey) => apiProvider.getTasks(groupKey);
 
-  Future<Null> addTask(
-      String taskName, String groupKey, int index, bool completed) async {
-    apiProvider.addTask(taskName, groupKey, index, completed);
+  Future<Null> addTask(String taskName, String groupKey) async {
+    apiProvider.addTask(taskName, groupKey);
   }
 
   Future<Null> updateTask(Task task) async {
@@ -76,11 +71,10 @@ class Repository {
   }
 
   //Subtasks
-  Future getSubtasks(String taskKey) => apiProvider.getSubtasks(taskKey);
+  Future getSubtasks(Task task) => apiProvider.getSubtasks(task);
 
-  Future<Null> addSubtask(
-      String taskKey, String subtaskName, int index, bool completed) async {
-    apiProvider.addSubtask(taskKey, subtaskName, index, completed);
+  Future<Null> addSubtask(String taskKey, String subtaskName) async {
+    apiProvider.addSubtask(taskKey, subtaskName);
   }
 
   Future<Null> updateSubtask(Subtask subtask) async {
@@ -91,7 +85,19 @@ class Repository {
     apiProvider.deleteSubtask(subtaskKey);
   }
 
+  ///Search For User
   Future searchUser(String searchTerm) => apiProvider.searchUser(searchTerm);
+
+  ///Group Members: Get, Post, Delete
+  Future<List<GroupMember>> getUsersAssignedToSubtask(
+          String subtaskKey) async =>
+      await apiProvider.getUsersAssignedToSubtask(subtaskKey);
+
+  Future assignSubtaskToUser(String subtaskKey, String username) =>
+      apiProvider.assignSubtaskToUser(subtaskKey, username);
+
+  Future<void> unassignSubtaskToUser(String subtaskKey, String username) =>
+      apiProvider.unassignSubtaskToUser(subtaskKey, username);
 }
 
 final repository = Repository();
